resolve comments: resume src/mock.rs, improve style
